Create SimulationV3
The simulation runs entirely in a web worker, the Simulation component
interacts with the web worker to seed it with some cells and keep an
history of each state computed by the simulation.

The Simulation component keeps a buffer of these states which are drawn
by the Canvas component.

For now modifying the attraction table doesn't work because it is
coupled to the underlying engine object